Add user ID parameter to Get method
The Get method in CharacterController and CharacterService, as well as its declaration in ICharacterService, now takes a user ID as a parameter. This change was made to ensure that only the authorized user's characters are fetched from the database. A claim has been added to get the user ID of the authenticated user. Now, when calling GetAllCharacters method, it only fetches the characters that belongs to the authenticated user.