Refactor GetById method to use IHttpContextAccessor
The Access UserID method in CharacterController.cs was refactored to use IHttpContextAccessor, removing the need to pass UserID from Controllers to Services. This modification improves the decoupling of components and simplifies the code. IHttpContextAccessor was added as a new dependency injection in CharacterService and registered in Program.cs. A new note was introduced to explain this approach.