Add user relationship to Character and update list method
This revision introduces the association between User and Character objects within the CharacterService and updates the service's method for listing characters. Now, when creating a new character, each character's User is set to be the currently logged in user. The service's method for listing characters now filters out characters not associated with the current user. Additionally, a new notes file is added which compares asynchronous code in ASP.net and Spring, detailing prevalent use in the former and also explains what servlets are.