Add Weapon entity and one-to-one relationship between Weapon and Character
This commit introduces a new Weapon entity which has properties of 'Id', 'Name', 'Damage', 'CharacterId' and 'Character' and also includes this new entity in 'DataContext' class. A unique one-to-one relationship is created between Weapon and Character respectively. The 'Character' class is updated to include this 'Weapon' entity. The migration for these changes is also included. This new entity and relationship allow us to assign a unique weapon to each character in the game, improving the role-playing features.