Add weapon functionality and update related files
Added the weapon functionality to the RPG game application, with corresponding Dtos, Controller methods, and Service methods. This change includes a new Weapon Controller, Weapon Service, and Dtos for getting and adding a weapon. Also, the Character Service was updated to include the Character's weapon data in the returned object. The AutoMapper Profile was extended to map the new weapon data as well. Moreover, small changes have been made in the AuthRepository for simplifying code and other updating other related files as per the new feature requirement. The new functionality allows adding a weapon to a character which enhances the battle feature in the RPG game.