Add specific exceptions and refactor services with Repositories
Introduced character and weapon repositories to encapsulate data access logic. Replaced direct DataContext usage in WeaponService with repository methods to respect separation of concerns.
Added CharacterNotFoundException to be thrown when character is not found instead of the generic Exception. This provides a clearer indication of the actual error that occurred. Also registered these new repositories in the DI container.