Remove 'AllowAnonymous' attribute and fix null reference issue
The 'AllowAnonymous' attribute was removed from the GetCharacterById method in the CharacterController to ensure that the User is always populated for each API request. This is to enhance the security feature of our system by always requiring an authenticated User while accessing this API.

In the CharacterRepository, a potential null reference exception was addressed. Instead of calling 'c.User!.Id', 'c.UserId' is used to avoid possible null reference issues. This ensures that the code is more robust and less error-prone.