Add Skill model and many-to-many relationship with Character model
In this commit, a Skill model was added to represent the skills a character has in the RPG game. Additionally, a many-to-many relationship between the Skill and Character models was established to cater for a character having many skills and a skill being held by many characters. This involved modification of the Character class to include a collection of Skills and creation of a join entity "CharacterSkill". Changes were also made in the DataContext class, migration files, and relevant notes.