Refactor test methods to use static import
Replaced direct usage of TestHelper methods in CharacterControllerTest and WeaponControllerTest with static import for brevity and improved code readability. Previous implementation required repetition of 'TestHelper' before each method invocation which made the test methods unnecessarily verbose. This change aims to simplify the test code and make it easier to read and maintain.