Update GitHub Actions workflow to run tests
The build job in the .github/workflows/dotnet.yml was renamed to 'test'. The working directory for the entire job was updated to 'dotnet-rpg.Tests', enabling all subsequent steps to use this directory without needing to specify it individually. This change allows for better clarity and easier maintenance of the workflow configuration.