Update dotnet version to environment variable in GitHub workflow
In the .github/workflows/dotnet.yml file, the .NET version used in the GitHub Actions workflow was hardcoded. This has been updated by setting an environment variable for the .NET version. This way, we only need to change the version in one place when there is a version update, thereby improving maintainability. The usage of the variable has been updated in the setup, test running, and result uploading steps of the workflow.