Add SkillService and tests, and refactor existing services and tests
In this commit, A new `skillService` and `skillServiceTests` classes were added to handle all skill related functionalities, as a necessary step for the domain driven design. The `CharacterService` was modified to not depend on the skills functionality.
Skill adding functionality from `CharacterController` has been moved to the newly created `SkillController`.
Similar changes were made in `CharacterServiceTests`, `WeaponServiceTests`, and a new class `CharacterLookupService` was created.
This was made mainly to adhere to the single responsibility principle and to simplify the unit tests. Further changes to note include the removal of the 'AddSkillToCharacter' method from CharacterService because the responsibility of adding a Skill to a Character was given to the newly created `SkillService`.
An architecture decision record (ADR) was also added to document our reasons behind these changes.