"Add fights, victories, defeats stats to Character model and service methods"
The commit introduces the tracking of statistics for characters in the RPG game, specifically their number of fights, victories, and defeats. This is achieved through additions in the Character model and corresponding migrations, updating the DbContext snapshot as well.

In addition, the FightService was extended with a new weapon attack method, which considers these new character stats. This required some additional lookup methods for the characters, introduced in the CharacterLookupService. The Fight controller was updated to include a post endpoint for these new attacks.

Moreover, the related Dtos were extended to cover the new properties, which ensures that these new stats can be transferred correctly between server and client.

Finally, some typos in the service naming were fixed for better consistency and readability.