Refactor fight service and introduce IAttackDto interface
This commit introduces a new interface, IAttackDto, which standardizes the necessary properties for performing an attack. By using this interface, we were able to refactor FightService to replace duplicated code within the WeaponAttack and SkillAttack methods with a single, reusable method, PerformAttack. This method uses a passed-in attack strategy method to further increase code reusability. The WeaponAttackDto and SkillAttackDto classes have been updated to implement this new interface.