Add AttackService to manage attack logic
A new service, AttackService, was implemented to encapsulate all the game attack logic. Before this change, the attack logic was intermixed within the FightService class which was violating the Single Responsibility Principle. Now, FightService is less complex and adheres to the said principle better. AttackService is also initialized in ApplicationBuilderExtensions, making it available anywhere necessary in the application. A minor adjustment was also done in FightLogger's damage logging, now considering zero damage as a case of no damage dealt.