Refactor FightService unit tests and enhance exception handling
Updated the FightService unit tests to test for service behavior rather than the internal functionality of each method. Additionally, new mocks were introduced for the recently added services, in order to stub their behavior during testing. This approach makes it easier to maintain the tests when the implementation details of the methods change.

Exceptions inside service methods now bubble up and are handled by the calling test, allowing for better test data cleanup and a more explicit display of error messages. This results in easier diagnosis of test failures. This is part of our ongoing effort to make our tests more resilient and easier to maintain.