Refactor FightServiceTests with generic SetupForAttack method
In our FightServiceTests, there was a lot of duplicated mocking setup across different test methods. To reduce this duplication and improve test maintainability, we have introduced a new generic method, SetupForAttack, that can be used in different test scenarios. This method centralizes the mocking setup and takes an AttackType parameter to control the specific mocking behavior. The result is a more concise and readable test code.