Refactor FightService test files and add clarification comment
The FightService test files have been moved into a new 'FightService' directory in order to improve file organisation and make the codebase easier to navigate. A comment has also been added in the FightServiceTests.cs file to explain why an exception is thrown, specifically for the purpose of terminating the fighting loop during tests when a character does not have weapon or skill.