"Expand AttackPerformService tests and refactor"
In this commit, we expanded the unit tests coverage for the AttackPerformService. This includes tests checking for proper handling of an attack on an already defeated opponent, the accurate update of fight statistics during a battle, and also verifying results after an attack.

Beyond testing, the AttackPerformService itself is refactored to enhance code readability including the renaming of BuildAttackResult to BuildAttackResultDto method.