Add ISpecification interface implementation to the ADR doc
In the ADR/2. Highscore - Specification Pattern.md, the ISpecification interface implementation was missing. I included the C# code for the interface with the properties `Criteria`, `Includes`, and `OrderBy`. This gives clearer insight into how our specification pattern is structured. The `IncludeStrings` property, although part of the interface, is not used in this particular implementation.